+++ lib/charms/opensearch/v0/opensearch_secrets.py
@@ -116,7 +116,7 @@ class OpenSearchSecrets(Object, RelationDataStore):
         # Leader has to maintain TLS and Dashboards relation credentials
         if not is_leader and label_key == CertType.APP_ADMIN.val:
             self._charm.tls.store_new_tls_resources(CertType.APP_ADMIN, event.secret.get_content())
-            if self._charm.is_tls_fully_configured():
+            if self._charm.tls.is_fully_configured():
                 self._charm.peers_data.put(Scope.UNIT, "tls_configured", True)
 
         elif is_leader and label_key == self._charm.secrets.password_key(KibanaserverUser):
